requirements.txt
streamlit==1.33.0
openai==1.23.2
python-dotenv==1.0.1
PyPDF2==3.0.1
tiktoken==0.5.2
pdfplumber==0.10.3
